+++ packages/video_player/video_player/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.12.0
+
+* Passes `backBufferDurationMs` from `VideoPlayerOptions` to the underlying platform interface.
+* Updates minimum supported SDK version to Flutter 3.38/Dart 3.10.
+
 ## 2.11.1
 
 * Optimizes caption retrieval with binary search.